+++ packages/ai/core/generate-text/stream-text.test.ts
@@ -3257,4 +3257,40 @@ describe('streamText', () => {
       ]);
     });
   });
+
+  describe('options.transform', () => {
+    it('should transform the stream', async () => {
+      const result = streamText({
+        model: new MockLanguageModelV1({
+          doStream: async () => ({
+            stream: convertArrayToReadableStream([
+              { type: 'text-delta', textDelta: 'Hello' },
+              { type: 'text-delta', textDelta: ', ' },
+              { type: 'text-delta', textDelta: `world!` },
+              {
+                type: 'finish',
+                finishReason: 'stop',
+                logprobs: undefined,
+                usage: { completionTokens: 10, promptTokens: 3 },
+              },
+            ]),
+            rawCall: { rawPrompt: 'prompt', rawSettings: {} },
+          }),
+        }),
+        experimental_transform: new TransformStream({
+          transform(chunk, controller) {
+            if (chunk.type === 'text-delta') {
+              chunk.textDelta = chunk.textDelta.toUpperCase();
+            }
+            controller.enqueue(chunk);
+          },
+        }),
+        prompt: 'test-input',
+      });
+
+      expect(
+        await convertAsyncIterableToArray(result.textStream),
+      ).toStrictEqual(['HELLO', ', ', 'WORLD!']);
+    });
+  });
 });