+++ packages/ai/core/index.ts
@@ -2,6 +2,7 @@ export { jsonSchema } from '@ai-sdk/ui-utils';
 export type { DeepPartial, Schema } from '@ai-sdk/ui-utils';
 export * from './data-stream';
 export * from './embed';
+export * from './generate-image';
 export * from './generate-object';
 export * from './generate-text';
 export * from './middleware';