+++ packages/google/src/google-generative-ai-language-model.ts
@@ -3,16 +3,17 @@ import {
   LanguageModelV1CallWarning,
   LanguageModelV1FinishReason,
   LanguageModelV1StreamPart,
+  LanguageModelV1ProviderMetadata,
 } from '@ai-sdk/provider';
 import {
   FetchFunction,
   ParseResult,
+  Resolvable,
   combineHeaders,
   createEventSourceResponseHandler,
   createJsonResponseHandler,
   postJsonToApi,
   resolve,
-  Resolvable,
 } from '@ai-sdk/provider-utils';
 import { z } from 'zod';
 import { convertJSONSchemaToOpenAPISchema } from './convert-json-schema-to-openapi-schema';
@@ -118,6 +119,7 @@ export class GoogleGenerativeAILanguageModel implements LanguageModelV1 {
         const { tools, toolConfig, toolWarnings } = prepareTools(
           mode,
           this.settings.useSearchGrounding ?? false,
+          this.modelId.includes('gemini-2'),
         );
 
         return {
@@ -241,6 +243,12 @@ export class GoogleGenerativeAILanguageModel implements LanguageModelV1 {
       rawCall: { rawPrompt, rawSettings },
       rawResponse: { headers: responseHeaders },
       warnings,
+      providerMetadata: {
+        google: {
+          groundingMetadata: candidate.groundingMetadata ?? null,
+          safetyRatings: candidate.safetyRatings ?? null,
+        },
+      },
       request: { body },
     };
   }
@@ -275,6 +283,8 @@ export class GoogleGenerativeAILanguageModel implements LanguageModelV1 {
       promptTokens: Number.NaN,
       completionTokens: Number.NaN,
     };
+    let providerMetadata: LanguageModelV1ProviderMetadata | undefined =
+      undefined;
 
     const generateId = this.config.generateId;
     let hasToolCalls = false;
@@ -314,6 +324,13 @@ export class GoogleGenerativeAILanguageModel implements LanguageModelV1 {
                 finishReason: candidate.finishReason,
                 hasToolCalls,
               });
+
+              providerMetadata = {
+                google: {
+                  groundingMetadata: candidate.groundingMetadata ?? null,
+                  safetyRatings: candidate.safetyRatings ?? null,
+                },
+              };
             }
 
             const content = candidate.content;
@@ -359,7 +376,12 @@ export class GoogleGenerativeAILanguageModel implements LanguageModelV1 {
           },
 
           flush(controller) {
-            controller.enqueue({ type: 'finish', finishReason, usage });
+            controller.enqueue({
+              type: 'finish',
+              finishReason,
+              usage,
+              providerMetadata,
+            });
           },
         }),
       ),
@@ -423,22 +445,86 @@ const contentSchema = z.object({
   ),
 });
 
-// limited version of the schema, focussed on what is needed for the implementation
-// this approach limits breakages when the API changes and increases efficiency
+// https://ai.google.dev/gemini-api/docs/grounding
+// https://cloud.google.com/vertex-ai/generative-ai/docs/multimodal/ground-gemini#ground-to-search
+export const groundingMetadataSchema = z.object({
+  webSearchQueries: z.array(z.string()).nullish(),
+  retrievalQueries: z.array(z.string()).nullish(),
+  searchEntryPoint: z
+    .object({
+      renderedContent: z.string(),
+    })
+    .nullish(),
+  groundingChunks: z
+    .array(
+      z.object({
+        web: z
+          .object({
+            uri: z.string(),
+            title: z.string(),
+          })
+          .nullish(),
+        retrievedContext: z
+          .object({
+            uri: z.string(),
+            title: z.string(),
+          })
+          .nullish(),
+      }),
+    )
+    .nullish(),
+  groundingSupports: z
+    .array(
+      z.object({
+        segment: z.object({
+          startIndex: z.number().nullish(),
+          endIndex: z.number().nullish(),
+          text: z.string().nullish(),
+        }),
+        segment_text: z.string().nullish(),
+        groundingChunkIndices: z.array(z.number()).nullish(),
+        supportChunkIndices: z.array(z.number()).nullish(),
+        confidenceScores: z.array(z.number()).nullish(),
+        confidenceScore: z.array(z.number()).nullish(),
+      }),
+    )
+    .nullish(),
+  retrievalMetadata: z
+    .union([
+      z.object({
+        webDynamicRetrievalScore: z.number(),
+      }),
+      z.object({}),
+    ])
+    .nullish(),
+});
+
+// https://cloud.google.com/vertex-ai/generative-ai/docs/multimodal/configure-safety-filters
+export const safetyRatingSchema = z.object({
+  category: z.string(),
+  probability: z.string(),
+  probabilityScore: z.number().nullish(),
+  severity: z.string().nullish(),
+  severityScore: z.number().nullish(),
+  blocked: z.boolean().nullish(),
+});
+
 const responseSchema = z.object({
   candidates: z.array(
     z.object({
-      content: contentSchema.optional(),
-      finishReason: z.string().optional(),
+      content: contentSchema.nullish(),
+      finishReason: z.string().nullish(),
+      safetyRatings: z.array(safetyRatingSchema).nullish(),
+      groundingMetadata: groundingMetadataSchema.nullish(),
     }),
   ),
   usageMetadata: z
     .object({
-      promptTokenCount: z.number(),
+      promptTokenCount: z.number().nullish(),
       candidatesTokenCount: z.number().nullish(),
-      totalTokenCount: z.number(),
+      totalTokenCount: z.number().nullish(),
     })
-    .optional(),
+    .nullish(),
 });
 
 // limited version of the schema, focussed on what is needed for the implementation
@@ -447,16 +533,18 @@ const chunkSchema = z.object({
   candidates: z
     .array(
       z.object({
-        content: contentSchema.optional(),
-        finishReason: z.string().optional(),
+        content: contentSchema.nullish(),
+        finishReason: z.string().nullish(),
+        safetyRatings: z.array(safetyRatingSchema).nullish(),
+        groundingMetadata: groundingMetadataSchema.nullish(),
       }),
     )
     .nullish(),
   usageMetadata: z
     .object({
-      promptTokenCount: z.number(),
+      promptTokenCount: z.number().nullish(),
       candidatesTokenCount: z.number().nullish(),
-      totalTokenCount: z.number(),
+      totalTokenCount: z.number().nullish(),
     })
     .nullish(),
 });