+++ packages/ai/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,28 @@
 
 ### Patch Changes
 
+- Updated dependencies [e3fac3f]
+  - ai@4.0.16
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [cc16a83]
+- Updated dependencies [cc16a83]
+  - ai@4.0.15
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [09a9cab]
+  - ai@4.0.14
+
+## 0.0.1
+
+### Patch Changes
+
 - Updated dependencies [9f32213]
   - ai@4.0.13