+++ packages/openai-compatible/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ai-sdk/openai-compatible
 
+## 0.0.6
+
+### Patch Changes
+
+- Updated dependencies [09a9cab]
+  - @ai-sdk/provider@1.0.2
+  - @ai-sdk/provider-utils@2.0.4
+
+## 0.0.5
+
+### Patch Changes
+
+- e958996: fix (provider/openai-compatible): remove unused index property from validation
+
 ## 0.0.4
 
 ### Patch Changes