+++ packages/openai/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @ai-sdk/openai
 
+## 1.0.8
+
+### Patch Changes
+
+- 09a9cab: feat (ai/core): add experimental generateImage function
+- Updated dependencies [09a9cab]
+  - @ai-sdk/provider@1.0.2
+  - @ai-sdk/provider-utils@2.0.4
+
 ## 1.0.7
 
 ### Patch Changes