+++ packages/provider-utils/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/provider-utils
 
+## 2.0.4
+
+### Patch Changes
+
+- Updated dependencies [09a9cab]
+  - @ai-sdk/provider@1.0.2
+
 ## 2.0.3
 
 ### Patch Changes